Moved NacBar components around.
