+++ src/App.jsx
@@ -3,15 +3,13 @@ import { Route } from "react-router-dom";
 import { useOktaAuth } from "@okta/okta-react";
 import { SecureRoute, LoginCallback } from "@okta/okta-react";
 import LandingPage from "./components/landingPage/page/LandingPage";
-import ResponsiveDrawer from "./AppDrawerNew";
+import ResponsiveDrawer from "./components/navbar/AppDrawer";
 import Profile from "./Profile";
 import "./App.css";
-
 function App() {
   const { authState, authService } = useOktaAuth();
   const login = async () => authService.login("/");
   const logout = async () => authService.logout("/");
-
   return (
     <>
       {authState.isAuthenticated ? (
@@ -28,5 +26,4 @@ function App() {
     </>
   );
 }
-
 export default App;
