+++ src/components/navbar/Drawer.jsx
@@ -6,7 +6,7 @@ const TopDivider = () => {
   return (
     <>
       <List>
-        {["Inbox", "Starred", "Send email", "Drafts"].map((text, index) => (
+        {["Subscriptions", "Pings", "Messages", "Subscribe"].map((text, index) => (
           <ListItem button key={text}>
             <ListItemText primary={text} />
           </ListItem>
